extractor: Support compressed Abiword documents
Even though Abiword does not save document files compressed, it does
read compressed document files with either .zabw or .abw.gz extensions.
Some people do use that to compress their documents, and some of that
people expect their indexer to index the documents, and do file bugs
about it.

Now that we use input streams to read the document, we can just plug
a GZlibDecompressor converter in the reading pipeline if we suspect
the file is not plain uncompressed XML. And we need to do this, since
all extensions do result in the same application/x-abiword content type.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/303