indexer: Notify corruption through our own file
Commit c454af95cce snuck the handling of constraint errors as
database corruptions in an insufficient manner. This would indeed
make the indexer exit swiftly instead of making a scene with a
barrage of warnings, but the untouched second half of this handling
is what happens on the next indexer restart, which was so far
nothing, the indexer would restart again, and probably find the
same constraint error, since the database is not corrupted in a
way that it will fail on construction, just in a way LocalSearch
knows it lost consistency.

To fix this, make LocalSearch itself track database corruption
at runtime with an ancillary hidden file in the cache dir, and
check for it existing at startup like we do with natural CORRUPTED
errors happening from the database.

We may want to handle other error types the same way in the future,
so an additional ancillary file can fit the bill for them all.