indexer: Handle removable mounts as separate indexer instances
Rework the handling of removable devices. These will be no longer indexed
in the main database at $HOME, but instead these will be indexed in a
dedicated database at the device itself.

This database will have a decicated endpoint, opened at a subpath at
"/org/freedesktop/LocalSearch3/..." with the last component of the path
being the file path of the removable device endpoint, with non alphanumeric
characters encoded as "_%x" of the ascii codepoint.

Size and access speed is nowadays barely a concern with these devices.
It makes sense to store the supporting data in the devices themselves,
instead of keeping the data in the main database, and making it mandatory
for everyone to stay aware of tracker:available, having hardly testable
code to deal with the expiration of this data...

The related bits of the test suite had obviously to be adapted to this
substantial change in handling of removable devices.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/347
Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/121
Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/89