extractor: Prepare for relative URIs
Add a --root commandline argument that will be used by the indexer
when spawning instances for removable volumes, and use it to resolve
relative "file:" URIs.

This also means the GFile URI is no longer (maybe) a valid identifier
for files in removable volumes, and the relative URI should be used
instead, a new TrackerExtractInfo getter is added for that, and used
in the relevant places across the metadata extractors.