indexer: Use relative file: URIs in removable volumes
Using relative URIs is necessary as the mount path is a detail of the
current system and user. Using file: URIs relative to the mountpoint
will ensure the data is universally valid.

In order to handle this, add a TrackerMinerFS::root property, that
the indexer instance for a given removable volume uses to resolve
absolute file:/// URIs to file: URIs relative to the removable volume.

This is passed along to the TrackerExtractWatchdog, so that it can
use the --root argument when spawning the extractor process for the
removable volume.