indexer: Open databases from removable devices in a sandboxed process
Besides the content of the removable device itself being handled by the
sandboxed metadata extractor, the database files themselves should be
considered a possible attack vector as well, that it makes sense to have
some defense in depth for.

Make use of the great versatility of TinySPARQL, and handle this database
connection as a sandboxed subprocess that opens a D-Bus endpoint to the
database through a dedicated D-Bus connection, for the main process to
transparently handle as a SPARQL connection. The indexer instance for
the mountpoint in the main process still does the operations deemed safe
(snapshotting the filesystem structure), and it still calls another
dedicated sandboxed extractor instance to handle this metadata, just that
it happens across a TrackerSparqlConnection that is transparently managed
within this new sandboxed process.

The sandbox in this new helper process borrows a lot from the metadata
extractor sandbox, just that it allows write operations in the database
folder of the removable device, plus some /tmp locations that SQLite likes
to have handy.