indexer: Build GDBus boilerplate for org.freedesktop.Tracker3.Miner.Files
This interface is so far handled by TrackerMinerProxy in src/common, and
this object is too oriented to managing a single TrackerMiner instance to
be useful for a generalization where the TrackerApplication manages multiple
indexer instances.

Have a cleaner start by using GDBus generated skeleton for this interface.
This will be integrated directly in the TrackerApplication in later commits.