indexer: Switch to "notify" systemd service type and sd_notify()
Getting the indexer database up to a working state may be a potentially
time consuming operation, due to different factors:
- Performing database integrity checks if it found corruption in
  previous runs.
- Updating to a newer internal database format as commanded by more
  recent versions of TinySPARQL.
- Updating the FTS index, due to parser changes or user locale
  changes.

These are handled on database start, which is necessary to create the
D-Bus SPARQL endpoint, which the indexer process waits for in order to
request the D-Bus name(s). This is important so that users of LocalSearch
can assume that the endpoint is ready for queries as soon as the D-Bus
name is available, and even mandatory if the service is implicitly
started up via D-Bus requests.

But this runs at odds with systemd and a "Type=dbus" unit as we define
for the indexer, since systemd will wait for the D-Bus name to be
claimed, which in some circumstances may take too long for its taste.

To fix this, switch to "Type=notify" and our implementation of
sd_notify() in order to notify readiness to systemd in a side channel.
Systemd will move along and maybe restart the service if it eventually
failed, and clients can keep relying on everything being up when the
name is claimed in the session bus.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/415