indexer: Add 'active' boolean property to the indexer object
We follow ::started/::finished signals to handle indexer activity,
but only the latter means that. The former only says tracker_miner_start()
was called on the indexer, which is pretty much once on startup.

Add a boolean 'active' property, that we can follow in the GApplication
to perform this kinf of things.