indexer: Check and save configuration against file on startup/shutdown
Serialize and store the current configuration as a gvariant, and save it
as an ancillary file besides the database. This file will be also checked
on startup, so that the indexer can know whether there were configuration
changes happening while the process was not running.

This is necessary as some of the possible setting changes may result in
the indexer not noticing files that previously should not be indexed but
now should. This config check will drop our regular optimizations to
avoid re-reading directories whose mtime did not change, and detect any
pre-existing file that should be indexed per the new config.

This will be useful in two situations:
- If the user modifies the configuration while the indexer is not running,
  e.g. on a TTY. This is not terribly common, but still useful.
- Changes in setting defaults, in this case the user will obtain the
  config changes from a system update.

This will result in contents being reindexed as if from scratch, with
all the expected files becoming available in the database.