config: Index home directory recursively
The prior approach to only index pictures/videos/music/documents
XDG folders recursively falls short for users that don't use these
folders to organize themselves. We want to make desktop search
just as useful for them.

This is not a light change, and the decision has not beent taken
lightly. The focus during the several last release cycles have
been focused in making indexing have both a good performance and
a low memory footprint, we have been preparing for this for a
long time.

Some non-scientific numbers, from testing locally:

With the previous configuration:
- Index would contain 48238 files, 32746 with content extracted.
- It would take ~90 seconds to index fully from scratch
- Restart on an up-to-date database would take 1.3s until idle

With the new configuration:
- Index contains 178149 files, 62093 with content extracted
- It takes ~144 seconds to index fully from scratch
- Restart on an up-to-date database takes 3.0s until idle
- Migrating from the old config takes ~185 seconds

So this does not change substantially the big picture, from
scratch indexing still takes a low amount of minutes, and restarts
after still take a low amount of seconds. Numbers may be larger
for some, but this is just a setting default.

Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/336
Closes: https://gitlab.gnome.org/GNOME/localsearch/-/issues/106