indexer: Coalesce config directory updates with pending changes
Avoid doubly scheduling an index root for it to be checked by the
TrackerFileNotifier, and merge the requests if the same directory
is still pending in the queue.

This also means we have to wait for the TrackerFileNotifier to be
started before we make any index root "current" (and start querying
around it), so we also coalesce such updates during startup.