tests: Fix possible race condition in test_offline_config_changes
The `with` statements used to wait for database changes through a
notifier might indirectly start the indexer, and this might result
in indexed folder changes early in startup. This is slightly racy
as those might be missed.

To fix this, restart the indexer within those `with` statements,
so the notifier and dbus watch are already set up before the
indexer is shutdown and restarted.