extractor: Make PDF extractor use XMP TrackerResource helpers
Use the TrackerResource helpers, instead of applying the data ourselves.
There's a few differences in data treatment that required changes to
expected test output. Notably:

- xmp's "subject" does not get converted to nie:subject, but a nao:hasTag
- Multiple creators are possible now

While at it, fix a small glitch with "creation_date" vs "date" fields in
the merged data. It seems we used the latter to set the resource, but only
used the former if there was no XMP data, oops.