indexer: Cancel current enumerator if contained in deleted file
The file notifier may receive a DELETE monitor event for a directory
while it is currently enumerating the same dir, or a subdir of it.
Don't expect that GFileEnumerator will notice this, and ensure to
cancel the operation and move on to the next thing to do.

This fixes errors like:

  Subject '...' is not is not a nfo:FileDataObject, cannot have property `tracker:extractorHash'

Since the directory was still enumerated and finished while some data
was already deleted.