indexer: Set fuller RDF when finishing a directory
We wait to fully process a directory contents' before we "complete"
it in RDF. Even though the data should be there, it does not hurt
to insert more self-contained RDF data, i.e. ensure that the
resource is a nfo:FileDataObject so it can have the missing
tracker:extractorHash property.