indexer: Drop big queue threshold when deleting from queue
This mechanism was added in commit c04293986a trying to optimize
deletion of folders when there was a too large processing queue left.
This limit of 1000 elements was arbitrary, trying to ensure that each
directory deletion didn't have to check too many thousands of files.

Shortly after, the "high water" mechanism was added (commit eb4a083aa52)
that prevented this processing queue to get too high (worth 2 batches,
or 1600 files), and waiting for it to be reduced before trying to
queue up further elements.

This is roughly the same ballpark, so it's not worth to keep this extra
limit, and make directory deletions check the full queue again for
children.