tests: Delete unused functional test
This test_writeback_image_details was left unused but remaining
in the tree. This is still using pre-3.0 writeback methods, and
testing specifics about it (specifically, that writeback operations
happened in the right order).

This order is now up to the direct D-Bus calls done to the writeback
service, so it is not rocket science as it previously was.