indexer: Simplify TrackerIndexingTree internals
Having a tree structure for configured folders is overkill for the
amount of configured folders there regularly are. We can make things
simpler with just an array and some smart ordering. In this array,
folders with longer (and deepmost) URIs are stored first, so looking
up in order does always find the deepmost configured folder first.