seccomp: Allow sandboxed process to disable SHSTK
SHSTK is a new security feature that is part of Intel CET. When enabled
in permissive mode, if a process that is SHSTK-compatible loads a
library that is not SHSTK-compatible, then SHSTK gets disabled
dynamically by glibc by issuing an arch_prctl(ARCH_SHSTK_DISABLE)
syscall (see glibc's dl_cet_disable_cet here).

The problem is: localsearch-extractor-3 is SHSTK-compatible and loads
SHSTK-incompatible libraries, but configures seccomp in a way that
disallows this syscall, causing a fatal SIGSYS when glibc tries to
disable SHSTK.

Unconditionally allowing this arch_prctl does not impact security,
because either (1) SHSTK is in permissive mode and disabling SHSTK is
therefore explicitly allowed, or (2) SHSTK is in enforcement mode and in
that case glibc will have set ARCH_SHSTK_LOCK which instructs the kernel
to reject ARCH_SHSTK_DISABLE anyway.

Fixes https://gitlab.gnome.org/GNOME/localsearch/-/issues/419