extractor: Return more consistent ISO8601 datetime strings
In tracker_date_guess(), a GDateTime is used to validate the guessed
datetime formats. Use it too to generate a more consistent ISO8601
string, with "Z" instead of timezone info if UTC is used, and without
microseconds if these are not provided.

Since these strings leak to extractor output, this of course requires
changes to the test suite.