extractor: Do not propagate a freed error in PDF extractor
If PDF extraction gets an error, we would store it in an inner
GError variable that is g_autoptr managed, and propagate it with
a prefix, however this inner error would be already freed at the
time of accessing it.

Fix this by giving ownership of the error with g_steal_pointer().
And while at it avoid the unnecessary prefix, this is already known
to come from a PDF file.