extractor: Validate UTF8-ness of EXIF comments
The charset of the EXIF comment is usually expressed in the
comment string itself per gexiv2 behavior, however this may not
be the full truth, and images may have non-ascii announced as
ascii, e.g.:

  $ exiv2 ...jpg |grep comment
  Exif comment    : charset=Ascii �;

Dont go along with this, and validate the string to avoid
non-utf8 in the comment. Fixes D-Bus warnings, as the resulting
RDF would fail to be inserted, the broken SPARQL update be
included in the error report back to the indexer process, and
the invalid UTF8 be rejected when creating the GVariant for this
D-Bus signal emission.